orbit_protocols::postgres_wire::storage

Trait TableStorage

pub trait TableStorage: Send + Sync {

    // Required methods
    fn initialize<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shutdown<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn metrics<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StorageMetrics> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn begin_transaction<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<StorageTransaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn commit_transaction<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx: &'life1 StorageTransaction,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn rollback_transaction<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx: &'life1 StorageTransaction,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn store_table_schema<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        schema: &'life1 TableSchema,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_table_schema<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<TableSchema>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_table_schemas<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<TableSchema>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_table_schema<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn insert_row<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
        row: &'life2 HashMap<String, SqlValue>,
        tx: Option<&'life3 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn insert_rows<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
        rows: &'life2 [HashMap<String, SqlValue>],
        tx: Option<&'life3 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn get_table_data<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<HashMap<String, SqlValue>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_rows<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
        updates: &'life2 HashMap<String, SqlValue>,
        _condition: Option<Box<dyn Fn(&'life3 HashMap<String, SqlValue>) -> bool + Send + Sync>>,
        tx: Option<&'life4 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             'life4: 'async_trait;
    fn delete_rows<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
        _condition: Option<Box<dyn Fn(&'life2 HashMap<String, SqlValue>) -> bool + Send + Sync>>,
        tx: Option<&'life3 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn truncate_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn store_index<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        index: &'life1 IndexSchema,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_index<'life0, 'life1, 'async_trait>(
        &'life0 self,
        index_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<IndexSchema>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_table_indexes<'life0, 'life1, 'async_trait>(
        &'life0 self,
        table_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<IndexSchema>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn remove_index<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        index_name: &'life1 str,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn store_view<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        view: &'life1 ViewSchema,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_view<'life0, 'life1, 'async_trait>(
        &'life0 self,
        view_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<ViewSchema>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_views<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<ViewSchema>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_view<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        view_name: &'life1 str,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn store_schema<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        schema: &'life1 SchemaDefinition,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_schema<'life0, 'life1, 'async_trait>(
        &'life0 self,
        schema_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<SchemaDefinition>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_schemas<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<SchemaDefinition>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_schema<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        schema_name: &'life1 str,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn store_extension<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        extension: &'life1 ExtensionDefinition,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_extension<'life0, 'life1, 'async_trait>(
        &'life0 self,
        extension_name: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<ExtensionDefinition>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_extensions<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<ExtensionDefinition>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_extension<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        extension_name: &'life1 str,
        tx: Option<&'life2 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn store_setting<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        key: &'life1 str,
        value: &'life2 str,
        tx: Option<&'life3 StorageTransaction>,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn get_setting<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_settings<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<HashMap<String, String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn checkpoint<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn compact<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn storage_size<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ProtocolResult<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Abstraction for PostgreSQL table storage operations

Required Methods

fn initialize<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Initialize the storage backend

fn shutdown<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Shutdown the storage backend gracefully

fn metrics<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StorageMetrics> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get current storage metrics

fn begin_transaction<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<StorageTransaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Begin a new transaction

fn commit_transaction<'life0, 'life1, 'async_trait>( &'life0 self, tx: &'life1 StorageTransaction, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Commit a transaction

fn rollback_transaction<'life0, 'life1, 'async_trait>( &'life0 self, tx: &'life1 StorageTransaction, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Rollback a transaction

fn store_table_schema<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, schema: &'life1 TableSchema, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Store table schema definition

fn get_table_schema<'life0, 'life1, 'async_trait>( &'life0 self, table_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<TableSchema>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get table schema by name

fn list_table_schemas<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<TableSchema>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List all table schemas

fn remove_table_schema<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, table_name: &'life1 str, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Remove table schema

fn insert_row<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, table_name: &'life1 str, row: &'life2 HashMap<String, SqlValue>, tx: Option<&'life3 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Insert row data into a table

fn insert_rows<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, table_name: &'life1 str, rows: &'life2 [HashMap<String, SqlValue>], tx: Option<&'life3 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Insert multiple rows into a table

fn get_table_data<'life0, 'life1, 'async_trait>( &'life0 self, table_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<HashMap<String, SqlValue>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get all rows from a table

fn update_rows<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>( &'life0 self, table_name: &'life1 str, updates: &'life2 HashMap<String, SqlValue>, _condition: Option<Box<dyn Fn(&'life3 HashMap<String, SqlValue>) -> bool + Send + Sync>>, tx: Option<&'life4 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, 'life4: 'async_trait,

Update rows in a table based on a condition (for now, update all rows)

fn delete_rows<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, table_name: &'life1 str, _condition: Option<Box<dyn Fn(&'life2 HashMap<String, SqlValue>) -> bool + Send + Sync>>, tx: Option<&'life3 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Delete rows from a table based on a condition (for now, delete all rows)

fn truncate_table<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, table_name: &'life1 str, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Clear all data from a table (but keep schema)

fn store_index<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, index: &'life1 IndexSchema, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Store index definition

fn get_index<'life0, 'life1, 'async_trait>( &'life0 self, index_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<IndexSchema>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get index by name

fn list_table_indexes<'life0, 'life1, 'async_trait>( &'life0 self, table_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<IndexSchema>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

List all indexes for a table

fn remove_index<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, index_name: &'life1 str, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Remove index

fn store_view<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, view: &'life1 ViewSchema, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Store view definition

fn get_view<'life0, 'life1, 'async_trait>( &'life0 self, view_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<ViewSchema>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get view by name

fn list_views<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<ViewSchema>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List all views

fn remove_view<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, view_name: &'life1 str, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Remove view

fn store_schema<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, schema: &'life1 SchemaDefinition, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Store schema definition

fn get_schema<'life0, 'life1, 'async_trait>( &'life0 self, schema_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<SchemaDefinition>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get schema by name

fn list_schemas<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<SchemaDefinition>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List all schemas

fn remove_schema<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, schema_name: &'life1 str, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Remove schema

fn store_extension<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, extension: &'life1 ExtensionDefinition, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Store extension definition

fn get_extension<'life0, 'life1, 'async_trait>( &'life0 self, extension_name: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<ExtensionDefinition>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get extension by name

fn list_extensions<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Vec<ExtensionDefinition>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List all extensions

fn remove_extension<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, extension_name: &'life1 str, tx: Option<&'life2 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Remove extension

fn store_setting<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, key: &'life1 str, value: &'life2 str, tx: Option<&'life3 StorageTransaction>, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Store configuration setting

fn get_setting<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str, ) -> Pin<Box<dyn Future<Output = ProtocolResult<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get configuration setting

fn list_settings<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<HashMap<String, String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List all settings

fn checkpoint<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Force a checkpoint (flush all pending data to disk)

fn compact<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Perform compaction (LSM-tree specific)

fn storage_size<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ProtocolResult<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Estimate storage size in bytes